LPS-136111 Move Tokenizer to data-engine-js-components-web
