+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/utils/Token.js
@@ -29,5 +29,3 @@ Token.LITERAL = 'Literal';
 Token.OPERATOR = 'Operator';
 Token.RIGHT_PARENTHESIS = 'Right Parenthesis';
 Token.VARIABLE = 'Variable';
-
-export default Token;
