+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/utils/rulesSupport.js
@@ -12,9 +12,8 @@
  * details.
  */
 
-import {Tokenizer} from 'dynamic-data-mapping-form-builder';
-
 import {getFieldProperty} from '../core/utils/fields';
+import {Tokenizer} from './Tokenizer';
 import {PagesVisitor} from './visitors.es';
 
 const DEFAULT_FIELD_NAMES_REGEX_FOR_EXPRESSION = /[a-zA-Z]+\d{8}/g;
