+++ modules/apps/data-engine/data-engine-js-components-web/test/js/utils/Tokenizer.js
@@ -12,8 +12,8 @@
  * details.
  */
 
-import Token from '../../../src/main/resources/META-INF/resources/js/expressions/Token.es';
-import Tokenizer from '../../../src/main/resources/META-INF/resources/js/expressions/Tokenizer.es';
+import {Token} from '../../../src/main/resources/META-INF/resources/js/utils/Token';
+import {Tokenizer} from '../../../src/main/resources/META-INF/resources/js/utils/Tokenizer';
 
 describe('Tokenizer', () => {
 	it('tokenizes single digit expressions', () => {
