+++ modules/apps/data-engine/data-engine-taglib/test/js/data-layout-builder/components/field-sets/FieldSetList.es.js
@@ -14,7 +14,6 @@
 
 import {ClayModalProvider} from '@clayui/modal';
 import {act, cleanup, fireEvent, render} from '@testing-library/react';
-import * as DDMForm from 'dynamic-data-mapping-form-builder';
 import React from 'react';
 import {DndProvider} from 'react-dnd';
 import {HTML5Backend} from 'react-dnd-html5-backend';
@@ -67,34 +66,6 @@ describe('FieldSets', () => {
 			name: 'Field53354166',
 			pages: FORM_VIEW.pages,
 		});
-
-		ddmFormSpy = jest
-			.spyOn(DDMForm, 'default')
-			.mockImplementation((props) => {
-				const state = {
-					...dataLayoutBuilder,
-					dispose: jest.fn(),
-					emit: jest.fn(),
-					formBuilderWithLayoutProvider: {
-						refs: {
-							layoutProvider: {
-								getRules: jest
-									.fn()
-									.mockImplementation(() => []),
-								on: jest.fn().mockImplementation(() => ({
-									removeListener: jest.fn(),
-								})),
-							},
-						},
-					},
-				};
-
-				props.layoutProviderProps.onLoad(state);
-
-				return state;
-			});
-
-		jest.useFakeTimers();
 
 		spySuccessToast = jest
 			.spyOn(toast, 'successToast')
